backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Remit Ai Backend code written in python"
readme = "README.md"
requires-python = "<3.14, >=3.10"
dependencies = [
    "crewai>=1.5.0",
    "crewai-tools>=1.5.0",
    "fastapi[all]>=0.121.2",
    "ollama>=0.6.1",
    "pydantic-settings",
    "python-dotenv",
    "python-levenshtein>=0.27.3",
    "requests>=2.31.0",
    "thefuzz[speedup]>=0.22.1",
]

[dependency-groups]
dev = [
    "pre-commit>=4.4.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.14.5",
]
[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = [
    "tests",
]
asyncio_mode = "auto"
